tools/bundlectl/requirements.txt
PyYAML>=6.0
PyNaCl>=1.5
zstandard>=0.22
pytest>=8.0
